Blacklists genemodders from Ascent.
Spawning as Ascent with genemodder tails/wings/ears doesn't remove them, so you can get a Gyne or Alate that has a vulpkanin tail. To prevent this, stops genemodders from spawning as Ascent. Feel free to unblacklist genemodders from Ascent if you code in a way to remove the genemodder cosmetics upon spawning as Ascent.